Move storage-adapter and get-stored-state to reducers/services
