+++ src/core/reducers/index.js
@@ -2,8 +2,8 @@ import { combineReducers } from 'redux';
 import { persistReducer } from 'redux-persist';
 import fuzzySettingsReducer from './fuzzy-settings';
 import generalSettingsReducer from './settings';
-import storage from '../store/storage-adapter';
-import getStoredState from './get-stored-state';
+import storage from './services/storage-adapter';
+import getStoredState from './services/get-stored-state';
 
 const persistConfig = {
   key: 'root',
