+++ src/core/reducers/services/get-stored-state.js
