+++ src/core/reducers/services/storage-adapter.js
